Added a "Rating" field to the model and Razor Page and updated Create/Delete/Details/Edit.
Notes: To re-seed the DB to include the Rating field, delete all the records in the table.

ALTERNATE METHOD listed in the tutorial is shown below; it involves deleting the table. This method did not go smoothly; the table was not automatically recreated and Update-Table did not work.

RECOVERY STEPS: found in https://stackoverflow.com/questions/34267155/how-do-i-force-drop-and-re-create-a-a-selected-table-in-code-first-migration

1  Remove the entity from the context (remember to remove all relationships with others entities)
2  Generate a Migration using Add-Migration (it will generate a drop table method)
3  Execute Migration using Update-Database
4  Add the entity to the context
5  Generate a new Migration using Add-Migration (it will generate a create table method)
6  Execute Migration using Update-Database

By modifying the database name in appsettings.json, a new database was generated. RazorPagesMovieContext-5678

ALTERNATE METHOD AS EXPLAINED IN TUTORIAL
Another option is to delete the database and use migrations to re-create the database. To delete the database in SSOX:

    Select the database in SSOX.

    Right click on the database, and select Delete.

    Check Close existing connections.

    Select OK.

    In the PMC, update the database:
    PowerShell

Update-Databas